Add link to Gitlab on how to report bugs in development banner
We want beta testers to report their findings, thus ensure the link to
Gitlab issues page is prominent. Also remove the string from translations,
since translations can't have links and anyway the string is not relevant
for translations and actual end users.

Also make Fevermap logo a link to fevermap.net and add the link as text
on the about page as well. Not pretty, but I leave it to UI designers
to iterate upon.