Bugfix: Partially revert c770e3ede733c1f so local dev env works again
The local development environment had stopped working. The API container
exited with error message:

  uwsgi: unrecognized option '--module=fevermap.wsgi:application'
         getopt_long() error

This is because UWSGIPLUGINLINE always evaluated to empty and in effect
made that the necessary '--plugins=python37' was always missing.

Re-implement this in some way that uses a properly named environment
variable and not just the first argument of the command.

# Conflicts:
#	api/entrypoint.sh